migrate interviewer commands
